pyproject.toml
[tool.poetry]
name = "mediamaster-v2"
version = "0.1.0"
description = "Multi-platform social media agent with modular architecture"
authors = ["Cyril Olivieri <cyril@olivieri.dev>"]
readme = "README.md"
packages = [{include = "mediamasterv2"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9"
pydantic-settings = "^2.5"
httpx = "^0.27"
tenacity = "^9.0"
PyYAML = "^6.0"
youtube-upload = "^0.9"
google-api-python-client = "^2.140"
discord-py = "^2.4"
python-telegram-bot = "^21"
twitchAPI = "^4"
pinterest-sdk = "^2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.24"
pytest-cov = "^5.0"
pytest-mock = "^3.14"
pytest-httpx = "^0.30"
ruff = "^0.6"
mypy = "^1.11"
pre-commit = "^4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
